Updated Readme for new functionality
I have updated the Readme file to properly explain how the updated SQL Select Query assembly works and how to pass the WorkflowID to the assembly. I have highlighted the case sensitive nature of the process field and requirement of a process field with a value of WorkflowID=Value and noted the error messages that will be returned in the batch when the requirement is not met.